Move classes which are implementation detail to internal
