+++ org.restlet.ext.openapi/src/main/java/org/restlet/ext/openapi/internal/RestletOpenApiContextBuilder.java
@@ -1,13 +1,13 @@
-/**
- * Copyright 2005-2026 Qlik
+/*
+ *  Copyright 2005-2026 Qlik
  *
- * The contents of this file is subject to the terms of the Apache 2.0 open source license available at
- * http://www.opensource.org/licenses/apache-2.0
+ *  The contents of this file is subject to the terms of the Apache 2.0 open
+ *  source license available at http://www.opensource.org/licenses/apache-2.0
  *
- * Restlet is a registered trademark of QlikTech International AB.
+ *  Restlet is a registered trademark of QlikTech International AB.
  */
 
-package org.restlet.ext.openapi;
+package org.restlet.ext.openapi.internal;
 
 import io.swagger.v3.oas.integration.GenericOpenApiContextBuilder;
 import io.swagger.v3.oas.integration.OpenApiConfigurationException;
